agents/key_master/requirements.txt
google-cloud-aiplatform
google-generativeai
uvicorn
fastapi
google-adk[a2a]
